🐛 Move edit cells to apollo-utils and extract shared functionality into apollo-common
