+++ .changeset/fresh-days-bow.md
@@ -1,0 +1,7 @@
+---
+'apollo-common': patch
+'@equinor/apollo-components': patch
+'@equinor/apollo-utils': patch
+---
+
+Move cells to @equinor/apollo-utils and extract common functionality into separate package
