+++ .changeset/perfect-mails-live.md
@@ -1,0 +1,5 @@
+---
+'@equinor/apollo-components': patch
+---
+
+Move react-hook-form to peer deps
