+++ packages/apollo-common/CHANGELOG.md
@@ -1,0 +1,14 @@
+# @equinor/apollo-utils
+
+## 0.1.1
+
+### Patch Changes
+
+- f9e695e: Add validateAsync utility to support async refinements
+
+## 0.1.0
+
+### Minor Changes
+
+- 8abf3eb: Initial version of Apollo Utils with jotai-form
+- 8abf3eb: Create apollo-utils package and jotai form example app
