+++ packages/apollo-common/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './PopoverCell'
+export * from './TypographyCustom'
