+++ packages/apollo-common/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "tsconfig/vite.json",
+  "include": ["src", "setupTest.ts"],
+  "compilerOptions": { "types": ["vitest/importMeta"] }
+}
