+++ packages/apollo-components/src/cells/index.ts
@@ -1,15 +1,7 @@
 export * from './ChipsCell'
 export * from './DynamicCell'
-export * from './EditableCheckboxCell'
-export * from './EditableDateCell'
-export * from './EditableDropdownCell'
-export * from './EditableNumberCell'
-export * from './EditableTextAreaCell'
-export * from './EditableTextFieldCell'
 export * from './HeaderCell'
 export * from './HierarchyCell'
-export * from './PopoverCell'
 export * from './SelectColumnDef'
 export * from './StickyCell'
-export * from './TypographyCustom'
 export * from './utils'
