+++ packages/apollo-utils/src/edit-cells/EditableTextFieldCell.tsx
@@ -1,9 +1,9 @@
 import { TextField } from '@equinor/eds-core-react'
 import { CellContext } from '@tanstack/react-table'
+import { TypographyCustom } from 'apollo-common'
 import { Controller, FieldPath, useFormContext } from 'react-hook-form'
 import styled from 'styled-components'
-import { FormMeta, useEditMode } from '../formMeta'
-import { TypographyCustom } from './TypographyCustom'
+import { FormMeta, useEditMode } from '../form-meta'
 import { getHelperTextProps } from './utils'
 
 export function EditableTextFieldCell<T extends FormMeta>(context: CellContext<T, unknown>) {
