+++ packages/apollo-utils/src/edit-cells/index.ts
@@ -1,0 +1,6 @@
+export * from './EditableCheckboxCell'
+export * from './EditableDateCell'
+export * from './EditableDropdownCell'
+export * from './EditableNumberCell'
+export * from './EditableTextAreaCell'
+export * from './EditableTextFieldCell'
